优化代码,满足前端显示

+++ yudao-module-fx/yudao-module-fx-biz/src/main/java/cn/iocoder/yudao/module/fx/controller/admin/skucostprice/vo/SkuCostpricePageReqVO.java
@@ -1,11 +1,13 @@
 package cn.iocoder.yudao.module.fx.controller.admin.skucostprice.vo;
 
-import lombok.*;
-import java.util.*;
-import io.swagger.v3.oas.annotations.media.Schema;
 import cn.iocoder.yudao.framework.common.pojo.PageParam;
-import java.math.BigDecimal;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
 import org.springframework.format.annotation.DateTimeFormat;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import static cn.iocoder.yudao.framework.common.util.date.DateUtils.FORMAT_YEAR_MONTH_DAY_HOUR_MINUTE_SECOND;
@@ -29,10 +31,10 @@ public class SkuCostpricePageReqVO extends PageParam {
     private String value;
 
     @Schema(description = "品类", example = "2")
-    private BigDecimal type;
+    private Integer type;
 
     @Schema(description = "属性", example = "8")
-    private BigDecimal paid;
+    private Integer paid;
 
     @Schema(description = "财务结算价", example = "25310")
     private BigDecimal costPrice;
